broke app up into single class modules
